polish(pu): polish eval_benchmark
